[Feat] Support mtaext files in mta read operations
The extensions are sent in parameter "-e" or "--extensions".
Currently, don't perform validations on the mtaext files except
unmarshaling and during the merge itself. Use the partial result if errors
occurs during the merge and return the errors in a "messages" property
on the result.
When using mtaext files, the hash is not returned to ensure the write
operations are performed on the data from the mta.yaml alone.

Also: move function readFile to file_ops and remove its duplications
and rename some variables used in commands.